VisualSpec

Allows the experimental mass spectra interactive visualization using the "plotly" graphical interface. Moreover, spectral peak-picking is performed. VisualSpec can be used to determine monoisotopic peak coordinates (m/z and Intensity) for a specific lipid sum compositions.

VisualSpec(path, mzmin = 0, mzmax = 0, span = 25)

Arguments

path
File path leading to the Tab Separated Value (TSV) text file (.txt) containing the m/z and Intensity values of each point of the experimental mass spectrum.

mzmin
The lower boundary of the spectrum m/z range. mzmin=0 is set as default.

mzmax
The upper boundary of the spectrum m/z range. mzmax=0 is set as default. Specifically, when mzmax=0, the full experimental m/z range is visualized. Conversely, if the user is interested in a specific spectral m/z range, the mzmin and mzmax need to be specified.

span
The "span" parameter value required by the "pick.peaks" function. This function is included in the "LIPIC" package, but belongs to the "ChemometricsWithR" package.

Value

A data frame containing the m/z and Intensity values related to peaks detected in the mass spectrum by the "pick.peaks" function.

References

Ron Wehrens: Chemometrics with R. Springer Verlag, Berlin Hei-delberg (2020). DOI: 10.1007/978-3-642-17841-2.

Carson Sievert: "Interactive Web-Based Data Visualization with R, plotly, and shiny" Chapman and Hall/CRC, 2020, https://plotly-r.com.

Examples