iLipic

iLipic is a computational tool invoked by the "Lipic_It" function for the recursive correction of both Type I and Type II isotopic effects.

iLipic(data, z, Res, t = 0.001, span = 25, ppm = 10, xseq = 0.001)

Arguments

data
This parameter is automatically set by the "Lipic_It" function.

z
This parameter is automatically set by the "Lipic_It" function.

Res
This parameter is automatically set by the "Lipic_It" function.

t
This parameter is automatically set by the "Lipic_It" function.

span
This parameter is automatically set by the "Lipic_It" function.

ppm
This parameter is automatically set by the "Lipic_It" function.

xseq
This parameter is automatically set by the "Lipic_It" function.

Value

A list of elements useful to build the Lipic_It output.

References

Andrea Castellaneta, Ilario Losito, Davide Coniglio, Beniamino Leoni, Pietro Santamaria, Maria A. Di Noia, Luigi Palmieri, Cosima D. Calvano, Tommaso R.I. Cataldi: "LIPIC: an automated workflow to account for isotopologues-related interferences in electrospray ionization high resolution mass spectra of phospholipids", Journal of the American Society for Mass Spectrometry, 2021. DOI: https://doi.org/10.1021/jasms.1c00008.

Martin Loos, Christian Gerber, Francesco Corona, Juliane Hollender, Heinz Singer: "Accelerated Isotope Fine Structure Calculation Using Pruned Transition Trees" Anal. Chem. 87, 5738–5744 (2015). DOI: https://doi.org/10.1021/acs.analchem.5b00941.

Ron Wehrens: Chemometrics with R. Springer Verlag, Berlin Hei-delberg (2020). DOI: https://doi.org/10.1007/978-3-642-17841-2.